Prevent frontend from crashing when user has broken state in localStorage;
If the database has been reset, the user will still have a reference
to the previously set current track in localStorage. Because the track
info stored in localStorage will most likely be garbage, it is better
to always fetch the current track from tracks returned from the backend
according to the track_id stored in currentTrack property of
localStorage. There will still be the wrong track loaded on start if the
database has been reset and a matching track_id happens to exist in the
tracks list returned by the backend. This should not be a big issue for
the user and thus not worth fixing (at least for now).