Query preview url from store logic;
** This commit removes the preloading feature from the UI **

This is preparation for bandcamp support.

Bandcamp requires the frontend to make an authorized request for the
preview url instead of using a constant url for the file. To be able
to support this, the application has to be able to execute backend logic
when ever the user requests a preview.