Fix track multiplication issue (label handling in user tracks query);
The previous version of the queryUserTracks SQL joined track__label to
the results which caused multiplication of the returned rows for a
single track. Because the track id was used in the UI as the key for the
react element, it would cause weird issues such as multiplication of
rows in the list.

This commit updates the query in a few ways. It cuts down the amount of
tracks for which details are built in order to enhance the performance.
It fixes the aforementioned issue by aggregating the labels and
returning a list of labels instead of a single label. This commit also
fixes how the labels are shown in the UI and how the artists and labels
are added to ignore as change in these was also required.